WORK IN PROGRESS: Attempt to map redis results in to Feathers
Complex to map the very different redis semantics into feathers.  This
commit is questionable, it is part of a work-in-progress that was
interrupted due to other priorities. Committing in case it is useful,
but it has problems with promises here. scanFrom should not be passed to
bluebird's try. Might be better reverting this and addressing the
previous state, or someone who has mapped different db semantics to
fetathers ignoring this fork entirely and starting from scratch.